filesystem/requirements.txt
fuse-python
requests
